moved user greeting, logout  func into ToDoColumn
